requirements.txt
alabaster==1.0.0
annotated-types==0.7.0
asgiref==3.8.1
babel==2.16.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.10
diff-match-patch==20241021
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-import-export==4.3.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
drf-yasg==1.21.8
fonttools==4.55.2
gunicorn==23.0.0
idna==3.10
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
murmurhash==1.0.11
neurokit2==0.2.10
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
preshed==3.0.9
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
sentry-sdk==2.19.2
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
smart-open==7.0.5
snowballstemmer==2.2.0
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.2
srsly==2.4.8
tablib==3.7.0
thinc==8.3.2
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.0
